+++ large_image/__init__.py
@@ -14,13 +14,17 @@
 #  limitations under the License.
 #############################################################################
 
-from pkg_resources import DistributionNotFound, get_distribution
-
 from . import tilesource  # noqa
 from .tilesource import canRead, getTileSource, open  # noqa
 
 try:
-    __version__ = get_distribution(__name__).version
-except DistributionNotFound:
+    from importlib.metadata import PackageNotFoundError
+    from importlib.metadata import version as _importlib_version
+except ImportError:
+    from importlib_metadata import PackageNotFoundError
+    from importlib_metadata import version as _importlib_version
+try:
+    __version__ = _importlib_version(__name__)
+except PackageNotFoundError:
     # package is not installed
     pass