+++ sources/tiff/large_image_source_tiff/__init__.py
@@ -24,7 +24,6 @@ import os
 import numpy
 import PIL.Image
 import tifftools
-from pkg_resources import DistributionNotFound, get_distribution
 
 from large_image import config
 from large_image.cache_util import LruCacheMetaclass, methodcache
@@ -37,8 +36,14 @@ from .tiff_reader import (InvalidOperationTiffException, IOTiffException,
                           ValidationTiffException)
 
 try:
-    __version__ = get_distribution(__name__).version
-except DistributionNotFound:
+    from importlib.metadata import PackageNotFoundError
+    from importlib.metadata import version as _importlib_version
+except ImportError:
+    from importlib_metadata import PackageNotFoundError
+    from importlib_metadata import version as _importlib_version
+try:
+    __version__ = _importlib_version(__name__)
+except PackageNotFoundError:
     # package is not installed
     pass