+++ utilities/converter/large_image_converter/__init__.py
@@ -14,7 +14,6 @@ from tempfile import TemporaryDirectory
 import numpy
 import psutil
 import tifftools
-from pkg_resources import DistributionNotFound, get_distribution
 
 import large_image
 
@@ -23,8 +22,14 @@ from . import format_aperio
 pyvips = None
 
 try:
-    __version__ = get_distribution(__name__).version
-except DistributionNotFound:
+    from importlib.metadata import PackageNotFoundError
+    from importlib.metadata import version as _importlib_version
+except ImportError:
+    from importlib_metadata import PackageNotFoundError
+    from importlib_metadata import version as _importlib_version
+try:
+    __version__ = _importlib_version(__name__)
+except PackageNotFoundError:
     # package is not installed
     pass